Partly fixed endpoints. Added Postman requests.

+++ src/main/java/org/kd/main/common/entities/Bank.java
@@ -13,7 +13,7 @@ public class Bank implements Serializable {
 
     @Id
     @GeneratedValue
-    private int id;
+    private long id;
 
     private String name;
     private String shortname;
@@ -21,7 +21,7 @@ public class Bank implements Serializable {
     private Bank() {
     }
 
-    public Bank(int id, String name, String shortname) {
+    public Bank(long id, String name, String shortname) {
         this(name, shortname);
         this.id = id;
     }
@@ -35,7 +35,7 @@ public class Bank implements Serializable {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
