+++ src/main/java/org/kd/main/common/entities/Transfer.java
@@ -10,27 +10,27 @@ public class Transfer implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
-    private int src_fund_id;
-    private int dest_fund_id;
+    private long id;
+    private long src_fund_id;
+    private long dest_fund_id;
     private float units;
     private boolean internal;
 
     private Transfer() {
     }
 
-    public Transfer(int src_fund_id, int dest_fund_id, float units, boolean internal) {
+    public Transfer(long src_fund_id, long dest_fund_id, float units, boolean internal) {
         this.src_fund_id = src_fund_id;
         this.dest_fund_id = dest_fund_id;
         this.units = units;
         this.internal = internal;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -47,7 +47,7 @@ public class Transfer implements Serializable {
         return "Transaction " + this.id + "from Customer " + getDest_fund_id() + " to party " + getSrc_fund_id();
     }
 
-    public int getDest_fund_id() {
+    public long getDest_fund_id() {
         return dest_fund_id;
     }
 
@@ -55,7 +55,7 @@ public class Transfer implements Serializable {
         this.dest_fund_id = dest_fund_id;
     }
 
-    public int getSrc_fund_id() {
+    public long getSrc_fund_id() {
         return src_fund_id;
     }
